+++ src/components/pages/home/all-activities/all-activities.tsx
@@ -1,35 +1,12 @@
 import { ActivityBasicDto } from "@/app/types/activity-schemas";
 import Image from "next/image";
 import Link from "next/link";
-import { useEffect, useState } from "react";
 
 interface AllActivitiesProps {
   activities: ActivityBasicDto[];
 }
 
 export default function AllActivities({ activities }: AllActivitiesProps) {
-  const [windowWidth, setWindowWidth] = useState<number>(0);
-
-  useEffect(() => {
-    const handleResize = () => {
-      setWindowWidth(window.innerWidth);
-    };
-
-    window.addEventListener("resize", handleResize);
-    setWindowWidth(window.innerWidth);
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
-
-  const displayedActivities =
-    windowWidth >= 1200
-      ? activities.slice(0, 8)
-      : windowWidth < 744
-        ? activities.slice(0, 4)
-        : activities;
-
   return (
     <ul
       className="grid grid-cols-4 gap-[2rem] w-[120rem] mb-[6rem]
@@ -37,7 +14,7 @@ export default function AllActivities({ activities }: AllActivitiesProps) {
     mobile:grid-cols-2 mobile:grid-rows-2 mobile:grid-auto-rows-[1fr]
     mobile:w-[38.8rem] mobile:mb-[4.6rem] mobile:px-[2rem] mobile:gap-[1.6rem]"
     >
-      {displayedActivities.map((activity) => (
+      {activities.map((activity) => (
         <li
           key={activity.id}
           className="rounded-[2rem] flex-column gap-[1.6rem]"
